+++ src/constants/commonErrorMessages.ts
@@ -3,3 +3,9 @@ export const COMMON_ERRORS = {
     USER_UPDATING_ERROR: 'An error occurred while updating details, Please try again !',
     VISITORS_COUNT_UPDATING_ERROR: 'An error occured while updating visitiors count, Please try again !'
 };
+
+export const LOGIN_ERROR_MESSAGE = {
+    INVALID_EMAIL_PASSWORD: 'Invalid Credentials !',
+    INTERNAL_SERVER_ERROR: 'Internal server error !',
+    INVALID_INPUT: 'Invalid input. Please provide a valid email/password !'
+};
