--- src/constants/errorMessages.ts
@@ -1,5 +1,0 @@
-export const LOGIN_ERROR_MESSAGE = {
-    INVALID_EMAIL_PASSWORD: 'Invalid Credentials !',
-    INTERNAL_SERVER_ERROR: 'Internal server error !',
-    INVALID_INPUT: 'Invalid input. Please provide a valid email/password !'
-};
