+++ src/controllers/admin/registrationController.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from 'express';
-import adminSignUpService from '../services/registration';
-import { ERRORS, ACCOUNT_MESSAGES } from '../constants/registrationMessages';
+import adminSignUpService from '../../services/admin/registerEmployeeByAdminService';
+import { ERRORS, ACCOUNT_MESSAGES } from '../../constants/registrationMessages';
 
 const randomPasswordGenerate = () => {
     return (Math.floor(Math.random() * 90000) + 10000) + '';
