+++ src/routes/adminRoutes.ts
@@ -1,15 +1,16 @@
 import express, { Router } from 'express';
-import registerController from '../controllers/registrationController';
-import loginController from '../controllers/loginController';
-import commonController from '../controllers/commonController';
+import registerEmployeeByAdminController from '../controllers/admin/registrationController';
+import getEmployeeDetailsByAdminController from '../controllers/admin/getEmployeeDetailsByAdminController';
+import updateEmployeeDetailsByAdminController from '../controllers/admin/updateEmployeeDetailsByAdminController';
+import commonController from '../controllers/common/commonController';
 import userSchema from '../middlewares/schemas/userSchema';
 import validateProfileRequest from '../middlewares/validateProfileUpdate';
 
 const adminRouter: Router = express.Router();
 
-adminRouter.post('/login', loginController.login);
-adminRouter.post('/registerEmployeeByAdmin', registerController.register);
-adminRouter.get('/getEmployeeDetailsByAdmin/:email', commonController.getUserDetails);
-adminRouter.put('/updateEmployeeDetailsByAdmin', validateProfileRequest(userSchema), commonController.updateProfile);
+adminRouter.post('/login', commonController.login);
+adminRouter.post('/registerEmployeeByAdmin', registerEmployeeByAdminController.register);
+adminRouter.get('/getEmployeeDetailsByAdmin/:email', getEmployeeDetailsByAdminController.getUserDetails);
+adminRouter.put('/updateEmployeeDetailsByAdmin', validateProfileRequest(userSchema), updateEmployeeDetailsByAdminController.updateProfile);
 
 export default adminRouter;
