+++ src/routes/commonRouter.ts
@@ -1,5 +1,5 @@
 import express, { Router } from 'express';
-import commonController from '../controllers/commonController';
+import commonController from '../controllers/common/commonController';
 
 const commonRouter: Router = express.Router();
